requirements/test.txt
# SHA1:69b8d3f48c3a28be7ece93019f2a9f4211037b4c
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
apipkg==1.5               # via execnet
attrs==20.2.0             # via pytest
coverage==5.3             # via pytest-cov
execnet==1.7.1            # via pytest-xdist
factory-boy==3.0.1        # via -r test.in
faker==4.1.3              # via factory-boy
iniconfig==1.0.1          # via pytest
more-itertools==8.5.0     # via pytest
packaging==20.4           # via pytest
pluggy==0.13.1            # via pytest
py==1.9.0                 # via pytest, pytest-forked
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.1        # via -r test.in
pytest-django==3.10.0     # via -r test.in
pytest-forked==1.3.0      # via pytest-xdist
pytest-mock==3.3.1        # via -r test.in
pytest-xdist==2.1.0       # via -r test.in
pytest==6.0.2             # via -r test.in, pytest-cov, pytest-django, pytest-forked, pytest-mock, pytest-xdist
python-dateutil==2.8.1    # via faker
six==1.15.0               # via packaging, python-dateutil
text-unidecode==1.3       # via faker
toml==0.10.1              # via pytest
